+++ app/src/main/java/com/example/polity_correct/ParliamentMember.java
@@ -4,7 +4,7 @@ import java.util.Date;
 
 public class ParliamentMember extends User{
 
-    public ParliamentMember(String userName, String password, String mail, Long yearOfBirth, int gander, UserType userType, String pg) {
+    public ParliamentMember(String userName, String password, String mail, Long yearOfBirth, long gander, UserType userType, String pg) {
         super(userName, password, mail, yearOfBirth, gander, userType, pg);
     }
